+++ demo/lib/demo/blog.ex
@@ -5,5 +5,7 @@ defmodule Demo.Blog do
     resource Demo.Blog.Author
     resource Demo.Blog.Comment
     resource Demo.Blog.Post
+    resource Demo.Blog.PostTag
+    resource Demo.Blog.Tag
   end
 end