+++ test/support/test_domain.ex
@@ -13,5 +13,8 @@ defmodule AshBackpex.TestDomain do
     resource(AshBackpex.TestDomain.AggregateItem)
     resource(AshBackpex.TestDomain.ReadOnlyEntry)
     resource AshBackpex.TestDomain.NonDefaultPrimaryKeyName
+    resource AshBackpex.TestDomain.ManyToManyPost
+    resource AshBackpex.TestDomain.ManyToManyCategory
+    resource AshBackpex.TestDomain.ManyToManyPostCategory
   end
 end